pyproject.toml
[tool.poetry]
name = "griddler"
version = "0.1.0"
description = ""
authors = ["Scott Olesen <ulp7@cdc.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
polars = "^0.20.31"
pytest = "^8.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
